refactor: move handlers into src/handlers/, make generic over Transport
Extracts run_message_router, run_1337_handler (and related helpers),
run_latency_handler, run_generic_command_handler, run_schedule_task
and config-watcher service from main.rs into focused submodules under
src/handlers/. Handler entry points now take <T: Transport, L: LoginCredentials>
so integration tests can substitute a fake transport. Moves PersonalBest
to handlers::tracker_1337, extracts config types to src/config.rs, and
hoists shared helpers (get_config_path) into util.
