+++ src/handlers/mod.rs
@@ -1,0 +1,5 @@
+pub mod commands;
+pub mod latency;
+pub mod router;
+pub mod schedules;
+pub mod tracker_1337;
