+++ src/handlers/router.rs
@@ -1,0 +1,24 @@
+use tokio::sync::{broadcast, mpsc::UnboundedReceiver};
+use tracing::{debug, info, instrument, trace};
+use twitch_irc::message::ServerMessage;
+
+/// Message router task that broadcasts incoming IRC messages to all handlers.
+///
+/// Reads from the twitch-irc receiver and broadcasts to all subscribed handlers.
+/// Exits when the incoming_messages channel is closed.
+#[instrument(skip(incoming_messages, broadcast_tx))]
+pub async fn run_message_router(
+    mut incoming_messages: UnboundedReceiver<ServerMessage>,
+    broadcast_tx: broadcast::Sender<ServerMessage>,
+) {
+    info!("Message router started");
+
+    while let Some(message) = incoming_messages.recv().await {
+        trace!(message = ?message, "Routing IRC message");
+
+        // Broadcast to all listeners (ignore errors if no receivers)
+        let _ = broadcast_tx.send(message);
+    }
+
+    debug!("Message router exited (connection closed)");
+}
